Squashed commit of the following:
commit ccf697bca700dffcff3c22f32a8ebba351db8ce3
Author: Taras Kushnir <kushnirTV@gmail.com>
Date:   Mon Apr 22 21:48:08 2019 -0700

    Cosmetic fixes

commit 3924497148e0c9490bd7f978b572edbb31828a57
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 17:04:06 2019 -0700

    Fix yaml

commit fd687b10a0473a2c42626d0bca6c5f770d4c93a4
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 17:02:55 2019 -0700

    Override for nightly

commit 96bc7c57b8947cf847d7333161b3421d478679cd
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 15:29:14 2019 -0700

    Build binary only in deploy branch

commit 54ecfc4c1ad2e6fa8b5276f600f2585d1a170c41
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 14:15:23 2019 -0700

    Rename branch to nightly in ci script [ci skip]

commit 7228ed16c3a2904317e9de158051a28195170fc3
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 14:14:39 2019 -0700

    Rename branch to nightly [ci skip]

commit ceabc6fcc24e54afba6067d0cd72b5d7db1e05be
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 14:12:33 2019 -0700

    Replace branch to nightly [ci skip]

commit 0eeb896fec77af71beed78f5928c1c9146c560c3
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 13:10:25 2019 -0700

    Do not create draft release

commit 00a44e5d33b215c9003218034b9a73e9cf8c848f
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 13:09:55 2019 -0700

    Do not create draft release [ci skip]

commit 26e0c17df0ac6ca9cafe5394fce08de44a15cf61
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 12:14:55 2019 -0700

    Fix artifact path

commit c5701dc5b320554372301a87ff98291283dbcf41
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 11:30:13 2019 -0700

    Deploy Xpiks in appveyor deploy branch

commit 47689a047f8fa8078a4ba4c07a9bdf8a1201bdfa
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 10:40:59 2019 -0700

    Add deploy instructions to appveyor script

commit 0a580b6164b8a64a5b9707fb94d3977bcedf916b
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Mon Apr 22 10:21:07 2019 -0700

    Disable linux tests in nightly [ci skip]

commit e3d4ab1ae22dd420a0f12764a6500861aeffc2b0
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Sat Apr 20 21:13:13 2019 -0700

    Cosmetic rename [ci skip]

commit d01ac7eb3256aaf61716725243e2d949c84b70de
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Sat Apr 20 20:24:23 2019 -0700

    Update .travis.yml

commit 36caf764f53d4f20592f9ab512bfaac922934f06
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 22:15:30 2019 -0700

    Fix path

commit 1527b1eda560a9e7b09879be8d6242c667a63192
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 21:52:44 2019 -0700

    Update .travis.yml

commit 8dfb0d1f9a27234c62b23daf8bae6ccfd8c5c776
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 21:51:58 2019 -0700

    Cosmetic fixes [ci skip]

commit 6d584bfa9b10cec5fee7d17176c30d653390416e
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 17:27:17 2019 -0700

    Update travis-ci-deploy-osx.sh

commit c91b40c772adcecee539ddb770e6448132b5f51b
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 17:05:35 2019 -0700

    Fix tar renaming

commit 09612259152205f86a77a40e1d85ef23396a2327
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 15:58:30 2019 -0700

    Download and extract exiftool before deploy

commit 0c7bfa67283cc83f95400d368d78ba384a8ff88c
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 15:31:05 2019 -0700

    Fix script permissions

commit fb1214fd6b861c03bd8abb3cf5af9214f242185e
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 15:05:44 2019 -0700

    Allow osx build in travis deploy

commit f032603ea0b89fe1299910592775601cd0fea921
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 15:02:22 2019 -0700

    Create travis-ci-deploy-linux.sh

commit e9eecb8705c741c1bea6bf445ed85ef908debb38
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 15:01:33 2019 -0700

    Add deploy section in travis.yml [ci skip]

commit 4eb648611e314aa1f952cecae44013c9e6da1718
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 14:57:08 2019 -0700

    Cleanup in debug mode [ci skip]

commit 155f425e2f9491a99419d046bb1512a2743d01f6
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 14:44:32 2019 -0700

    Create deploy script for macOS on travis-ci [ci skip]

commit d5a8b014effe450a0e3730baa081061b52b3e4aa
Author: Taras Kushnir <505555+ribtoks@users.noreply.github.com>
Date:   Fri Apr 19 13:24:52 2019 -0700

    Prepare deploy_mac script for travis deploy [ci skip]

+++ scripts/deploy/deploy_mac.sh
@@ -1,31 +1,45 @@
 #!/bin/bash
 
+# Can set variables 
+# $BUILD_MODE (debug/release), 
+# $BUILD_DIR - path to build directory with executable
+# $DEPLOY_TOOL - path to macdeployqt
+# $VERSION - default (1.5.2)
+# $APP_NAME - default (Xpiks)
+
 dir_resolve()
 {
     cd "$1" 2>/dev/null || return $?  # cd to desired directory; if fail, quell any error messages but return exit status
     echo "`pwd -P`" # output full, link-resolved path
 }
 
+BUILD_MODE="${BUILD_MODE:-release}"
+
 XPIKS_ROOT=$(git rev-parse --show-toplevel)
 
 XPIKS_DEPS_DIR="${XPIKS_ROOT}/src/xpiks-qt/deps"
 XPIKS_QT_DIR="${XPIKS_ROOT}/src/xpiks-qt"
-BUILD_DIR="${XPIKS_ROOT}/src/build-xpiks-qt-Desktop_Qt_5_6_2_clang_64bit-Release"
+
+DEFAULT_BUILD_DIR="${XPIKS_ROOT}/src/build-xpiks-qt-Desktop_Qt_5_6_2_clang_64bit-Release"
+BUILD_DIR="${BUILD_DIR:-$DEFAULT_BUILD_DIR}"
 
 QT_BIN_DIR=~/Qt5.6.2/5.6/clang_64/bin
-DEPLOY_TOOL="$QT_BIN_DIR/macdeployqt"
+DEFAULT_DEPLOY_TOOL="$QT_BIN_DIR/macdeployqt"
+DEPLOY_TOOL="${DEPLOY_TOOL:-$DEFAULT_DEPLOY_TOOL}"
 
-LIBS_PATH="${XPIKS_ROOT}/libs/release"
+LIBS_PATH="${XPIKS_ROOT}/libs/${BUILD_MODE}"
 
-APP_NAME=Xpiks
-VERSION="1.5.2"
+DEFAULT_VERSION="1.5.2"
+DEFAULT_APP_NAME="Xpiks"
 VOL_NAME="Xpiks"
+APP_NAME="${APP_NAME:-$DEFAULT_APP_NAME}"
+VERSION="${VERSION:-$DEFAULT_VERSION}"
 
 # ----------------------------------------
 
 if [ ! -f "${XPIKS_DEPS_DIR}/exiftool/exiftool" ]; then
     echo "Exiftool not found! Please put latest production release into the ../deps/ dir"
-    exit
+    exit 1
 fi
 
 DMG_BACKGROUND_IMG="dmg-background.jpg"
@@ -38,12 +52,12 @@ STAGING_DIR="./osx-release-staging"
 
 if [ ! -d "$BUILD_DIR" ]; then
     echo "Build directory not found: $BUILD_DIR"
-    exit
+    exit 1
 fi
 
 if [ -d "/Volumes/${VOL_NAME}" ]; then
     echo "Please unmount existing Volume ${VOL_NAME} first!"
-    exit
+    exit 1
 fi
 
 rm -v -rf "${STAGING_DIR}" "${DMG_TMP}" "${DMG_FINAL}"
@@ -84,6 +98,11 @@ pushd "$FRAMEWORKS_DIR"
 
 for lib in "${LIBS_TO_DEPLOY[@]}"
 do
+    if [ ! -f "$LIBS_PATH/$lib" ]; then
+        echo "Critical: $lib not found in $LIBS_PATH"
+        exit 1
+    fi
+
     echo "Processing $lib..."
     cp -v "$LIBS_PATH/$lib" .
 
@@ -98,6 +117,11 @@ done
 
 for lib in "${FFMPEG_LIBS[@]}"
 do
+    if [ ! -f "$LIBS_PATH/$lib" ]; then
+        echo "Dependent $lib not found in $LIBS_PATH"
+        continue
+    fi
+    
     echo "Copying $lib..."
     cp -v "$LIBS_PATH/$lib" .
     
@@ -150,7 +174,7 @@ SIZE=`echo "${SIZE} + 32.0" | bc | awk '{print int($1+0.5)}'`
 
 if [ $? -ne 0 ]; then
    echo "Error: Cannot compute size of staging dir"
-   exit
+   exit 1
 fi
 
 echo "Size is estimated to: $SIZE"
@@ -176,32 +200,34 @@ popd
 mkdir /Volumes/"${VOL_NAME}"/.background
 cp "${DMG_BACKGROUND_PATH}" /Volumes/"${VOL_NAME}"/.background/
 
-# tell the Finder to resize the window, set the background,
-#  change the icon size, place the icons in the right position, etc.
-echo '
-   tell application "Finder"
-     tell disk "'${VOL_NAME}'"
-           open
-           set current view of container window to icon view
-           set toolbar visible of container window to false
-           set statusbar visible of container window to false
-           set the bounds of container window to {480, 300, 980, 570}
-           set statusbar visible of container window to false
-           set position of every item to {600, 400}
-           set viewOptions to the icon view options of container window
-           set arrangement of viewOptions to not arranged
-           set icon size of viewOptions to 128
-           set background picture of viewOptions to file ".background:'${DMG_BACKGROUND_IMG}'"
-           set position of item "'${APP_NAME}'.app" of container window to {110, 135}
-           set position of item "Applications" of container window to {395, 130}
-           set name of item "Applications" to " "
-           close
-           open
-           update without registering applications
-           delay 2
-     end tell
-   end tell
-' | osascript
+if [ "$BUILD_MODE" == "release" ]; then
+    # tell the Finder to resize the window, set the background,
+    #  change the icon size, place the icons in the right position, etc.
+    echo '
+       tell application "Finder"
+         tell disk "'${VOL_NAME}'"
+               open
+               set current view of container window to icon view
+               set toolbar visible of container window to false
+               set statusbar visible of container window to false
+               set the bounds of container window to {480, 300, 980, 570}
+               set statusbar visible of container window to false
+               set position of every item to {600, 400}
+               set viewOptions to the icon view options of container window
+               set arrangement of viewOptions to not arranged
+               set icon size of viewOptions to 128
+               set background picture of viewOptions to file ".background:'${DMG_BACKGROUND_IMG}'"
+               set position of item "'${APP_NAME}'.app" of container window to {110, 135}
+               set position of item "Applications" of container window to {395, 130}
+               set name of item "Applications" to " "
+               close
+               open
+               update without registering applications
+               delay 2
+         end tell
+       end tell
+    ' | osascript
+fi
 
 sync
 
@@ -215,6 +241,8 @@ hdiutil convert "${DMG_TMP}" -format UDZO -imagekey zlib-level=9 -o "${DMG_FINAL
 shasum "${DMG_FINAL}"
 
 # clean up
-# rm -rf "${DMG_TMP}"
+if [ "$BUILD_MODE" == "debug" ]; then
+    rm -rvf "${DMG_TMP}"
+fi
 
 echo 'Done.'
